Module aws_sdk_voiceid::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

CreateDomain

Fluent builder constructing a request to CreateDomain.

DeleteDomain

Fluent builder constructing a request to DeleteDomain.

DeleteFraudster

Fluent builder constructing a request to DeleteFraudster.

DeleteSpeaker

Fluent builder constructing a request to DeleteSpeaker.

DescribeDomain

Fluent builder constructing a request to DescribeDomain.

DescribeFraudster

Fluent builder constructing a request to DescribeFraudster.

DescribeFraudsterRegistrationJob

Fluent builder constructing a request to DescribeFraudsterRegistrationJob.

DescribeSpeaker

Fluent builder constructing a request to DescribeSpeaker.

DescribeSpeakerEnrollmentJob

Fluent builder constructing a request to DescribeSpeakerEnrollmentJob.

EvaluateSession

Fluent builder constructing a request to EvaluateSession.

ListDomains

Fluent builder constructing a request to ListDomains.

ListFraudsterRegistrationJobs

Fluent builder constructing a request to ListFraudsterRegistrationJobs.

ListSpeakerEnrollmentJobs

Fluent builder constructing a request to ListSpeakerEnrollmentJobs.

ListSpeakers

Fluent builder constructing a request to ListSpeakers.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

OptOutSpeaker

Fluent builder constructing a request to OptOutSpeaker.

StartFraudsterRegistrationJob

Fluent builder constructing a request to StartFraudsterRegistrationJob.

StartSpeakerEnrollmentJob

Fluent builder constructing a request to StartSpeakerEnrollmentJob.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateDomain

Fluent builder constructing a request to UpdateDomain.