Module aws_sdk_voiceid::error

Errors that can occur when calling the service.

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

You do not have sufficient permissions to perform this action. Check the error message and try

ConflictException

The request failed due to a conflict. Check the ConflictType and error message for

CreateDomainError

Error type for the CreateDomain operation.

DeleteDomainError

Error type for the DeleteDomain operation.

DeleteFraudsterError

Error type for the DeleteFraudster operation.

DeleteSpeakerError

Error type for the DeleteSpeaker operation.

DescribeDomainError

Error type for the DescribeDomain operation.

DescribeFraudsterError

Error type for the DescribeFraudster operation.

DescribeFraudsterRegistrationJobError

Error type for the DescribeFraudsterRegistrationJob operation.

DescribeSpeakerEnrollmentJobError

Error type for the DescribeSpeakerEnrollmentJob operation.

DescribeSpeakerError

Error type for the DescribeSpeaker operation.

EvaluateSessionError

Error type for the EvaluateSession operation.

InternalServerException

The request failed due to an unknown error on the server side.

ListDomainsError

Error type for the ListDomains operation.

ListFraudsterRegistrationJobsError

Error type for the ListFraudsterRegistrationJobs operation.

ListSpeakerEnrollmentJobsError

Error type for the ListSpeakerEnrollmentJobs operation.

ListSpeakersError

Error type for the ListSpeakers operation.

ListTagsForResourceError

Error type for the ListTagsForResource operation.

OptOutSpeakerError

Error type for the OptOutSpeaker operation.

ResourceNotFoundException

The specified resource cannot be found. Check the ResourceType and error message for

ServiceQuotaExceededException

The request exceeded the service quota. Refer to Voice ID Service

StartFraudsterRegistrationJobError

Error type for the StartFraudsterRegistrationJob operation.

StartSpeakerEnrollmentJobError

Error type for the StartSpeakerEnrollmentJob operation.

TagResourceError

Error type for the TagResource operation.

ThrottlingException

The request was denied due to request throttling. Please slow down your request rate. Refer to

UntagResourceError

Error type for the UntagResource operation.

UpdateDomainError

Error type for the UpdateDomain operation.

ValidationException

The request failed one or more validations; check the error message for more details.

Enums

CreateDomainErrorKind

Types of errors that can occur for the CreateDomain operation.

DeleteDomainErrorKind

Types of errors that can occur for the DeleteDomain operation.

DeleteFraudsterErrorKind

Types of errors that can occur for the DeleteFraudster operation.

DeleteSpeakerErrorKind

Types of errors that can occur for the DeleteSpeaker operation.

DescribeDomainErrorKind

Types of errors that can occur for the DescribeDomain operation.

DescribeFraudsterErrorKind

Types of errors that can occur for the DescribeFraudster operation.

DescribeFraudsterRegistrationJobErrorKind

Types of errors that can occur for the DescribeFraudsterRegistrationJob operation.

DescribeSpeakerEnrollmentJobErrorKind

Types of errors that can occur for the DescribeSpeakerEnrollmentJob operation.

DescribeSpeakerErrorKind

Types of errors that can occur for the DescribeSpeaker operation.

EvaluateSessionErrorKind

Types of errors that can occur for the EvaluateSession operation.

ListDomainsErrorKind

Types of errors that can occur for the ListDomains operation.

ListFraudsterRegistrationJobsErrorKind

Types of errors that can occur for the ListFraudsterRegistrationJobs operation.

ListSpeakerEnrollmentJobsErrorKind

Types of errors that can occur for the ListSpeakerEnrollmentJobs operation.

ListSpeakersErrorKind

Types of errors that can occur for the ListSpeakers operation.

ListTagsForResourceErrorKind

Types of errors that can occur for the ListTagsForResource operation.

OptOutSpeakerErrorKind

Types of errors that can occur for the OptOutSpeaker operation.

StartFraudsterRegistrationJobErrorKind

Types of errors that can occur for the StartFraudsterRegistrationJob operation.

StartSpeakerEnrollmentJobErrorKind

Types of errors that can occur for the StartSpeakerEnrollmentJob operation.

TagResourceErrorKind

Types of errors that can occur for the TagResource operation.

UntagResourceErrorKind

Types of errors that can occur for the UntagResource operation.

UpdateDomainErrorKind

Types of errors that can occur for the UpdateDomain operation.