Module aws_sdk_voiceid::model

Data structures used by operation inputs/outputs.

Modules

authentication_configuration

See AuthenticationConfiguration

authentication_result

See AuthenticationResult

domain

See Domain

domain_summary

See DomainSummary

enrollment_config

See EnrollmentConfig

enrollment_job_fraud_detection_config

See EnrollmentJobFraudDetectionConfig

failure_details

See FailureDetails

fraud_detection_configuration

See FraudDetectionConfiguration

fraud_detection_result

See FraudDetectionResult

fraud_risk_details

See FraudRiskDetails

fraudster

See Fraudster

fraudster_registration_job

See FraudsterRegistrationJob

fraudster_registration_job_summary

See FraudsterRegistrationJobSummary

input_data_config

See InputDataConfig

job_progress

See JobProgress

known_fraudster_risk

See KnownFraudsterRisk

output_data_config

See OutputDataConfig

registration_config

See RegistrationConfig

server_side_encryption_configuration

See ServerSideEncryptionConfiguration

speaker

See Speaker

speaker_enrollment_job

See SpeakerEnrollmentJob

speaker_enrollment_job_summary

See SpeakerEnrollmentJobSummary

speaker_summary

See SpeakerSummary

tag

See Tag

Structs

AuthenticationConfiguration

The configuration used to authenticate a speaker during a session.

AuthenticationResult

The authentication result produced by Voice ID, processed against the current session state and streamed

Domain

Contains all the information about a domain.

DomainSummary

Contains a summary of information about a domain.

EnrollmentConfig

Contains configurations defining enrollment behavior for the batch job.

EnrollmentJobFraudDetectionConfig

The configuration defining the action to take when a speaker is flagged by the fraud detection system

FailureDetails

Contains error details for a failed batch job.

FraudDetectionConfiguration

The configuration used for performing fraud detection over a speaker during a session.

FraudDetectionResult

The fraud detection result produced by Voice ID, processed against the current session state and streamed

FraudRiskDetails

Details regarding various fraud risk analyses performed against the current session state and streamed

Fraudster

Contains all the information about a fraudster.

FraudsterRegistrationJob

Contains all the information about a fraudster registration job.

FraudsterRegistrationJobSummary

Contains a summary of information about a fraudster registration job.

InputDataConfig

The configuration containing input file information for a batch job.

JobProgress

Indicates the completion progress for a batch job.

KnownFraudsterRisk

Contains details produced as a result of performing known fraudster risk analysis on a speaker.

OutputDataConfig

The configuration containing output file information for a batch job.

RegistrationConfig

The configuration definining the action to take when a duplicate fraudster is detected, and the

ServerSideEncryptionConfiguration

The configuration containing information about the customer-managed KMS Key used for encrypting

Speaker

Contains all the information about a speaker.

SpeakerEnrollmentJob

Contains all the information about a speaker enrollment job.

SpeakerEnrollmentJobSummary

Contains a summary of information about a speaker enrollment job.

SpeakerSummary

Contains a summary of information about a speaker.

Tag

A tag that can be assigned to a Voice ID resource.

Enums

AuthenticationDecision

ConflictType

DomainStatus

DuplicateRegistrationAction

ExistingEnrollmentAction

FraudDetectionAction

FraudDetectionDecision

FraudDetectionReason

FraudsterRegistrationJobStatus

ResourceType

SpeakerEnrollmentJobStatus

SpeakerStatus

StreamingStatus